Uncompress 'zip' Archives

zip_extract() always restores modification times of the extracted files and directories.

zip_extract(
  zipfile,
  files = NULL,
  overwrite = TRUE,
  junkpaths = FALSE,
  exdir = "."
)

Arguments

zipfile
Path to the zip file to uncompress.

files
Character vector of files to extract from the archive. Files within directories can be specified, but they must use a forward slash as path separator, as this is what zip files use internally. If NULL (default), all files will be extracted.

overwrite
Whether to overwrite existing files. If FALSE and a file already exists, then an error is thrown.

junkpaths
Whether to ignore all directory paths when creating files. If TRUE, all files will be extracted to the top level of exdir.

exdir
Directory to uncompress the archive to. If it does not exist, it will be created.

Value

An fs::fs_path() character vecttor of extracted files, invisibly.

This is like the "internal" method of utils::unzip() and unlike zip::unzip() (See r-lib/zip#35). The zip_list() C function is called separately to get file names, which are not yet returned from the C internal used for extraction.

Permissions

If the zip archive stores permissions and was created on Unix, the permissions will be restored.

Examples

.old_wd <- setwd(tempdir())
library(fs)
dir_create("mydir")
cat("first file",  file = path("mydir", "file1"))
cat("second file", file = path("mydir", "file2"))

z <- file_temp(ext = "zip")
zip_create(z, files = "mydir")

## List contents
zip_list(z)

#> # A tibble: 3 x 5
#>   filename    compressed_size uncompressed_size timestamp           permissions
#>   <fs::path>      <fs::bytes>       <fs::bytes> <dttm>              <fs::perms>
#> 1 mydir                     0                 0 2021-02-08 05:02:08 rwxr-xr-x  
#> 2 mydir/file1              15                10 2021-02-08 05:02:08 rw-rw-rw-  
#> 3 mydir/file2              16                11 2021-02-08 05:02:08 rw-rw-rw-  

## Extract
tmp <- path_temp()
out <- zip_extract(z, exdir = tmp)
length(out)

#> [1] 3

setwd(.old_wd)