List information on files in a ZIP archive

zip_info() is similar to the unzip -l command. Compared to zip::zip_list(), the data frame is returned as a tibble when available and columns use human-readable classes like fs::dir_info().

zip_ls() returns only the names of the files in an archive, like fs::dir_ls().

zip_info(archive)

zip_ls(archive)

Arguments

archive
Path to an existing ZIP archive.

Value

A data.frame with 'fs' class columns.

Details

USes the options(fs.use_tibble = FALSE) to disable tibble printing.

Examples

zip_info(zip_example())

#> # A tibble: 4 x 5
#>   filename      compressed_size uncompressed_si… timestamp           permissions
#>   <fs::path>        <fs::bytes>      <fs::bytes> <dttm>              <fs::perms>
#> 1 example                     0                0 2019-02-23 21:48:06 rwxr-xr-x  
#> 2 example/dir                 0                0 2019-02-23 21:49:36 rwxr-xr-x  
#> 3 example/dir/…              11                6 2019-02-23 21:48:24 rw-r--r--  
#> 4 example/file1              11                6 2019-02-23 21:48:28 rw-r--r--  

zip_ls(zip_example())

#> example           example/dir       example/dir/file2 example/file1