Class for an external zip process

zip_process() returns an R6 class that represents a zip process. It is implemented as a subclass of processx::process.

zip_process()

Value

A zip_process R6 class object, a subclass of processx::process.

Using the zip_process class

zp <- zip_process()$new(zipfile, files, recurse = TRUE,
                         poll_connection = TRUE,
                         stderr = tempfile(), ...)

See processx::process for the class methods.

Arguments:

zipfile: Path to the zip file to create.

files: List of file to add to the archive. Each specified file or directory in is created as a top-level entry in the zip archive.

recurse: Whether to add the contents of directories recursively.

include_directories: Whether to explicitly include directories in the archive. Including directories might confuse MS Office when reading docx files, so set this to FALSE for creating them.

poll_connection: passed to the initialize method of processx::process, it allows using processx::poll() or the poll_io() method to poll for the completion of the process.

stderr: passed to the initialize method of processx::process, by default the standard error is written to a temporary file. This file can be used to diagnose errors if the process failed.

... passed to the initialize method of processx::process.

Examples

dir.create(tmp <- tempfile())
write.table(iris, file = file.path(tmp, "iris.ssv"))
zipfile <- tempfile(fileext = ".zip")
zp <- zip_process()$new(zipfile, tmp)

#> Error in rethrow_call(c_processx_exec, command, c(command, args), stdin,     stdout, stderr, pty, pty_options, connections, env, windows_verbatim_args,     windows_hide_window, private, cleanup, wd, encoding, paste0("PROCESSX_",         private$tree_id, "=YES")): cannot start processx process '' (system error 2, No such file or directory) @unix/processx.c:604 (processx_exec)

zp$wait()

#> Error in eval(expr, envir, enclos): object 'zp' not found

zp$get_exit_status()

#> Error in eval(expr, envir, enclos): object 'zp' not found

zip_list(zipfile)

#> Warning: path[1]="/var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T//RtmpKNxCz4/file84c5a96055b.zip": No such file or directory

#> Error in zip_list(zipfile): Cannot open zip file `(null)`